# lease: Persist remainingTTL to prevent indefinite auto-renewal of long lived leases