--- a/lease/leasepb/lease.proto
+++ b/lease/leasepb/lease.proto
@@ -13,6 +13,7 @@ option (gogoproto.goproto_enum_prefix_all) = false;
 message Lease {
   int64 ID = 1;
   int64 TTL = 2;
+  int64 RemainingTTL = 3;
 }
 
 message LeaseInternalRequest {