@jpbetz 2018-07-16
@gyuho Batching only briefly crossed my mind, but it's something we should clearly do. I'll add it shortly.