@jpbetz 2018-07-17
Due to the size of this PR, I'll split it into three commits:
- .proto change and resulting codegen
- Lessor config and logging change
- checkpointing mechanism