@xiang90 2018-07-17
The approach looks good to me. We need to have some benchmarks to show the overhead is acceptable in normal cases.