@jpbetz 2018-07-23
@xiang90 @gyuho 

Ran two benchmarks:

### Checkpoint heap size

Checked etcd server heap size up to 10,000,000 live leases.

- With checkpointing 3.3GB
- Without checkpointing 3.3GB
This makes sense given that the heap is a slice of structs that contain only three int64s, so the total memory usage for all the entries is only about 40MB.  I've removed the limit on this heap as it does not seem to be needed.

### Checkpoint rate limit

Set leases to checkpoint every 1s, created 15k of them, and then checked server performance with `benchmark put`.

- Without checkpointing - write latency ~ 0.006ms
- With checkpointing, maxLeaseCheckpointBatchSize=0 - write latency ~0.015ms
- With checkpointing, maxLeaseCheckpointBatchSize=1000 - write latency ~0.008ms
- With checkpointing, maxLeaseCheckpointBatchSize=1000, leaseCheckpointRate=10000 - write latency ~0.008ms
- With checkpointing, maxLeaseCheckpointBatchSize=1000, leaseCheckpointRate=1000 - write latency ~0.006ms

Since 1,000,000 checkpoints per sec seems sufficient, and the limits of maxLeaseCheckpointBatchSize=1000, leaseCheckpointRate=1000 appear to have negligible impact on performance, I've gone with those settings.