@gyuho 2018-07-23
> With checkpointing, maxLeaseCheckpointBatchSize=1000, leaseCheckpointRate=1000 - write latency ~0.006ms

Results look good to me. Thanks for benchmarks!