@gyuho 2018-07-23
@jpbetz Also, can you add this to CHANGELOG? Just separate commit or PR should be fine. Thanks.