@keeplowkey 2021-12-08
As a normal user, how do I use the "lease checkpointing" mechanism：
By upgrading the etcd server to certain version or setting some specific parammeters in config file？I don't know.
Much thx for your help~ @jpbetz @xiang90 @gyuho