@serathius 2021-12-08
We have recently found that Lease Checkpointing doesn't work as intended in https://github.com/etcd-io/etcd/pull/13491. Fix is planned to be released in v3.5.2. With this release you should be able to enable release checkpointing by providing `--experimental-enable-lease-checkpoint` and `--experimental-enable-lease-checkpoint-persist` flags