@keeplowkey 2021-12-08
> We have recently found that Lease Checkpointing doesn't work as intended in #13491. Fix is planned to be released in v3.5.2. With this release you should be able to enable release checkpointing by providing `--experimental-enable-lease-checkpoint` and `--experimental-enable-lease-checkpoint-persist` flags

@serathius If I want to enable lease checkpointing, just start etcd server with flag "--experimental-enable-lease-checkpoint true" will do the trick? Current etcd version: 3.4.7.  Looking forward to your reply.